+++ SecurityComplianceCenter/changelog.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## 1.1.1 (2020-04-28)
+
+- Fix: Set-SccLabelLocalization - A bad entry would prevent subsequent updates to a label in non-cached mode
+- Fix: Set-SccLabelLocalization - Tries to write too long localization strings
+
 ## 1.1.0 (2020-04-27)
 
 - New: Import-SccLabelLocalizationXml - imports the localization XML exported from the AIP portal